% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizeMatrixes.R
\name{cramerEtaMatrix}
\alias{cramerEtaMatrix}
\title{Effect size matrices}
\usage{
cramerEtaMatrix(data)
}
\description{
Create effect size matrices for all type of variables
Effect sizes - Cramer's V, ETA coefficient, Pearson's r
}
