% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBox.R
\name{plotBox}
\alias{plotBox}
\title{Plot box plots}
\usage{
plotBox(
  var,
  res,
  title = NA,
  varname = NA,
  resLevels = NA,
  seed = NA,
  legend = F
)
}
\arguments{
\item{var}{}

\item{res}{}
}
\description{
Plot box plots for continuous vs categorical variables.
}
