% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHB.R
\name{plotHB}
\alias{plotHB}
\title{Plot Histogram and Box plot in the same graph}
\usage{
plotHB(var, title = NA, varname = NA, seed = NA, histX = F)
}
\arguments{
\item{var}{The variable that needs to be plotted. The variable needs to be
a quantitative variable.}

\item{title}{The title of the plot}

\item{varname}{The variable name}

\item{seed}{The seed for the color of the plot}

\item{histX}{Logical value to specify whether to print the X axis for the
histogram}
}
\description{
This function plots the histogram and the box plot for
quantitative variables in the same plot.
}
