% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSc.R
\name{plotSc}
\alias{plotSc}
\title{Plot Scatter plots}
\usage{
plotSc(
  var,
  res,
  pName = NA,
  resName = NA,
  seed = NA,
  title = T,
  method = "auto",
  transX = NA,
  transY = NA,
  verb = T
)
}
\description{
Plot scatter plots for 2 continuous variables with or without transformations
}
