% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomColor.R
\name{randomColor}
\alias{randomColor}
\title{Get random colors}
\usage{
randomColor(seed = NA, count = 1)
}
\arguments{
\item{seed}{Seed for the random color. Use this to get the same color each time}

\item{count}{Number of colors that need to be returned. Defaults to 1.}
}
\value{
A vector of randomly generated colors.
}
\description{
This function gives random colors using the RGB function.
}
