% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R, R/plotBar2.R
\name{plotBar}
\alias{plotBar}
\alias{plotBar2}
\title{Plot bar charts}
\usage{
plotBar(
  var,
  res,
  group = NA,
  varname = NA,
  title = NA,
  varLevels = NA,
  resLevels = NA,
  groupLevels = NA,
  seed = NA,
  legend = T,
  number = T,
  type = "fill",
  verb = T
)

plotBar2(
  var,
  res,
  varname = NA,
  title = NA,
  varLevels = NA,
  resLevels = NA,
  seed = NA,
  legend = T,
  type = "fill",
  verb = T
)
}
\arguments{
\item{var}{}

\item{res}{}

\item{group}{}
}
\description{
Plot bar charts

Plot bar charts
}
\section{Functions}{
\itemize{
\item \code{plotBar()}: function plots stacked bar charts for 2 categorical variables.

\item \code{plotBar2()}: function plots stacked bar charts for 2 categorical variables.

}}
