% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSc.R
\name{plotSc}
\alias{plotSc}
\title{Plot Scatter plots}
\usage{
plotSc(
  var,
  res,
  pName = NA,
  resName = NA,
  seed = NA,
  title = NA,
  titleT = T,
  method = "auto",
  transX = NA,
  transY = NA,
  verbT = T,
  labXT = T,
  labYT = T,
  r2T = T
)
}
\description{
Plot scatter plots for 2 continuous variables with or without transformations
}
